Update code to match PSR-1, PSR-2 and PSR-4
